refactor: migrate scripts/schema_validator.py to src/quality/ (#344)
T4 of ADR-0010 implementation. Moves schema_validator to src/quality/
and updates caller:
- tests/test_quality_system.py

Refs #344
