+++ tests/test_quality_system.py
@@ -16,10 +16,10 @@ from pathlib import Path
 # Add scripts directory to path
 sys.path.insert(0, str(Path(__file__).parent.parent / "scripts"))
 
-from schema_validator import validate_front_matter
 from skills_manager import SkillsManager
 
 from src.quality.agent_reviewer import review_agent_output
+from src.quality.schema_validator import validate_front_matter
 
 
 def test_issue_15_prevention():
